requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
environs==9.5.0
idna==3.3
marshmallow==3.15.0
packaging==21.3
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
secure-smtplib==0.1.1
soupsieve==2.3.2.post1
urllib3==1.26.9
